requirements.txt
python-telegram-bot
cairosvg
svglib
requests